Karolynbranch

+++ public/assets/js/index.js
@@ -55,8 +55,8 @@ $(document).ready(function () {
     const projectList = $("#projectList")
     var projectInfo = [];
 
-    $(document).on("click", "button.delete", handleProjectDelete)
-    $(document).on("click", "button.edit", handleProjectEdit)
+    $(document).on("click", "#projectDelete", handleProjectDelete)
+    $(document).on("click", "#projectEdit", handleProjectEdit)
     $(document).on("click", "#projectTitle", projectDetails)
 
     let url = window.location.search;
@@ -95,6 +95,7 @@ $(document).ready(function () {
         })
     }
 
+
     function deleteProject(id) {
         $.ajax({
             method: "DELETE",
@@ -105,29 +106,28 @@ $(document).ready(function () {
     }
 
     function createRow(project) {
-        var formattedDate = new Date(project.createdAt);
-        //formattedDate = moment(formattedDate).format("MMMM Do YYYY, h:mm:ss a");
-        let updatedLast = new Date(project.updatedAt);
-        //updatedLast= moment(updatedLast).format("MMMM Do YYYY, h:mm:ss a");
 
         let projectCard = projectList.append(
             `<div class="card" id="${project.title}"><div class="card-header">
-            <h3 id="projectTitle"><a href="#">${project.title}</a></h3><button class="delete btn btn-danger" id="projectDelete"><i class="fas fa-trash delete-project"></i></button>
+            <h3 id="projectTitle"><a href="#">${project.title}</a></h3>
             <button class="edit btn btn-info" id="projectEdit"><i class="fas fa-edit edit-project"></i></button>
-            `
+            <button class="delete btn btn-danger" id="projectDelete"><i class="fas fa-trash delete-project"></i></button>
+            </div>
+            </div>`
+
         )
         projectCard.data("project", project)
 
         return projectCard;
 
     }
-    
+
     function displayEmpty() {
-        
+
         projectList.empty();
 
         projectList.append(`<p>You have no projects to display</p>`)
-        
+
     }
 
     function handleProjectEdit() {
@@ -137,61 +137,84 @@ $(document).ready(function () {
             .parent()
             .parent()
             .data("project")
+        console.log(currentProject)
         window.location.href = "/add?project_id=" + currentProject.projectId;
     }
 
     function handleProjectDelete() {
-        let currentProject = $(this)
-            .parent().parent().parent().data("project")
-        console.log(currentProject)
-        deleteProject(currentProject.projectId)
+        let answer = window.confirm("Are you sure you would like to delete this project?")
+
+        if (answer) {
+            let currentProject = $(this)
+                .parent().parent().parent().data("project")
+            console.log(currentProject)
+            deleteProject(currentProject.projectId)
+        }
+        else { return; }
     }
-    
-    function projectDetails () {
+
+    function projectDetails() {
         $(".current-project-details").empty()
+        $("#task-section").empty()
         var selectedProject = event.target.innerHTML;
-        for(var i = 0;i < projectInfo.length; i++){
-            if(projectInfo[i].title == selectedProject){
+        for (var i = 0; i < projectInfo.length; i++) {
+            if (projectInfo[i].title == selectedProject) {
                 selectedProject = projectInfo[i].projectId;
                 selectedProject = selectedProject - 1;
             }
         }
+
         let currentProject = projectInfo[selectedProject];
+        let team = currentProject.team;
+        team = JSON.parse(team)
 
         $(".current-project-details").append(`
             <h3 class="selected-project">${currentProject.title}</h3>
             <p class="project-status"><strong>Status: </strong><span class="current-project-status">${currentProject.status}</span></p>
-            <p class="project-assignees"><strong>Assignees: </strong><span class="current-project-assignees">${currentProject.team}</span></p>
+            <p class="project-assignees"><strong>Assignees: </strong><span class="current-project-assignees"></span></p>
             <p class="project-budget"><strong>Budget (in hours): </strong><span class="current-project-budget">${currentProject.budget}</span></p>
             <p><strong>Description:</strong></p>
             <p class="current-project-desc"> ${currentProject.description}</p>
+            <small>Created on: ${currentProject.createdAt}</small>
+            <small>Last updated: ${currentProject.updatedAt}</small>
         `)
+
+        team.forEach(member => {
+            $(".current-project-assignees").append(member + ", ")
+        });
+
+        $("#task-section").append(`
+        <h2 class="col-md-10">Task</h2><button class="add-task" id="addTaskBtn" style= "float : right;"data-toggle="modal" data-target="#taskModal"><i class="fas fa-plus"></i> Add Task</button>`)
+
         getTasks(currentProject.projectId);
     }
     var tasks;
     var name;
 
     function getTasks(id) {
+        $("#task-display").empty();
+
         $.when(
             $.get("/api/Task", function (data) {
                 tasks = data;
             })
-        ).done( function() {
-            for (var i =0;i<tasks.length;i++){
-                if(tasks[i].projectId == id) {
-                    $(".current-project-details").append(`
-                    <li class="task-item">
-                        <p>
-                            <input type="checkbox" class="completed-task">
-                            <span class="task-title">${tasks[i].taskName}</span><br/>
-                            <strong>Assigned to: </strong><span class="task-assignee">${tasks[i].User.firstname} ${tasks[i].User.lastname}</span><br/>
-                        </p>
-                    </li>
+        ).done(function () {
+            for (var i = 0; i < tasks.length; i++) {
+                if (tasks[i].projectId == id) {
+                    $("#task-display").append(`
+                    <div class = "col-md-3"><div class = "card" id ="taskCard">
+                <div class = "card-header" id = "task-header"><h5 style="color:white;">${tasks[i].taskName}</h5></div><br>
+                <strong>Assigned to: ${tasks[i].User.firstname} ${tasks[i].User.lastname}</strong>
+                <details><p>${tasks[i].taskDescription}</p>
+                <p>Created on: ${tasks[i].createdAt}</p>
+                </div></div></details> <br>          
+        
                     `)
                 }
             }
         })
     }
+
 
     getProject();
 });
